AG-6869: Cleanup CartesianSeriesV2.
